+++ packages/babel-preset-expo/src/use-dom-directive-plugin.ts
@@ -110,28 +110,18 @@ export function expoUseDomDirectivePlugin(
         if (isProduction) {
           // MUST MATCH THE EXPORT COMMAND!
           const hash = crypto.createHash('sha1').update(outputKey).digest('hex');
-          const outputName = `www.bundle/${hash}.html`;
-
-          if (platform === 'ios') {
-            proxyModule.push(`const source = { uri: "${outputName}" };`);
-          } else if (platform === 'android') {
-            proxyModule.push(`const source = { uri: "file:///android_asset/${outputName}" };`);
-          } else {
-            throw new Error(
-              'production "use dom" directive is not supported yet for platform: ' + platform
-            );
-          }
+          proxyModule.push(`const filePath = "${hash}.html";`);
         } else {
           proxyModule.push(
             // Add the basename to improve the Safari debug preview option.
-            `const source = { uri: new URL("/_expo/@dom/${fileBasename}?file=" + ${JSON.stringify(outputKey)}, require("react-native/Libraries/Core/Devtools/getDevServer")().url).toString() };`
+            `const filePath = "${fileBasename}?file=" + ${JSON.stringify(outputKey)};`
           );
         }
 
         proxyModule.push(
           `
 export default React.forwardRef((props, ref) => {
-  return React.createElement(WebView, { ref, ...props, source });
+  return React.createElement(WebView, { ref, ...props, filePath });
 });`
         );