+++ packages/@expo/cli/src/api/graphql/client.ts
@@ -2,16 +2,15 @@ import {
   cacheExchange,
   Client,
   CombinedError as GraphqlError,
+  AnyVariables,
+  DocumentInput,
   createClient as createUrqlClient,
-  dedupExchange,
   fetchExchange,
   OperationContext,
   OperationResult,
-  PromisifiedSource,
-  TypedDocumentNode,
+  OperationResultSource,
 } from '@urql/core';
 import { retryExchange } from '@urql/exchange-retry';
-import { DocumentNode } from 'graphql';
 
 import * as Log from '../../log';
 import { fetch } from '../../utils/fetch';
@@ -31,7 +30,6 @@ type SessionHeaders = {
 export const graphqlClient = createUrqlClient({
   url: getExpoApiBaseUrl() + '/graphql',
   exchanges: [
-    dedupExchange,
     cacheExchange,
     retryExchange({
       maxDelayMs: 4000,
@@ -65,11 +63,11 @@ export const graphqlClient = createUrqlClient({
 
 /* Please specify additionalTypenames in your Graphql queries */
 export interface StricterClient extends Client {
-  query<Data = any, Variables extends object = object>(
-    query: DocumentNode | TypedDocumentNode<Data, Variables> | string,
-    variables: Variables | undefined,
+  query<Data = any, Variables extends AnyVariables = AnyVariables>(
+    query: DocumentInput<Data, Variables>,
+    variables: Variables,
     context: Partial<OperationContext> & { additionalTypenames: string[] }
-  ): PromisifiedSource<OperationResult<Data, Variables>>;
+  ): OperationResultSource<OperationResult<Data, Variables>>;
 }
 
 export async function withErrorHandlingAsync<T>(promise: Promise<OperationResult<T>>): Promise<T> {