+++ packages/expo-router/src/layouts/withLayoutContext.tsx
@@ -72,7 +72,10 @@ export function useFilterScreenChildren(
   }, [children]);
 }
 
-/** Return a navigator that automatically injects matched routes and renders nothing when there are no children. Return type with children prop optional */
+/**
+ * Returns a navigator that automatically injects matched routes and renders nothing when there are no children.
+ * Return type with `children` prop optional.
+ */
 export function withLayoutContext<
   TOptions extends object,
   T extends ComponentType<any>,
@@ -83,13 +86,9 @@ export function withLayoutContext<
   processor?: (
     options: ScreenProps<TOptions, TState, TEventMap>[]
   ) => ScreenProps<TOptions, TState, TEventMap>[]
-): ForwardRefExoticComponent<
-  PropsWithoutRef<PickPartial<ComponentProps<T>, 'children'>> & RefAttributes<unknown>
-> & {
-  Screen: (props: ScreenProps<TOptions, TState, TEventMap>) => null;
-} {
-  const Navigator = forwardRef<unknown, PropsWithoutRef<ComponentProps<T>>>(
-    ({ children: userDefinedChildren, ...props }, ref) => {
+) {
+  return Object.assign(
+    forwardRef(({ children: userDefinedChildren, ...props }: any, ref) => {
       const contextKey = useContextKey();
 
       const { screens } = useFilterScreenChildren(userDefinedChildren, {
@@ -105,13 +104,14 @@ export function withLayoutContext<
         return null;
       }
 
-      // @ts-expect-error
       return <Nav {...props} id={contextKey} ref={ref} children={sorted} />;
+    }),
+    {
+      Screen,
     }
-  );
-
-  // @ts-expect-error
-  Navigator.Screen = Screen;
-  // @ts-expect-error
-  return Navigator;
+  ) as ForwardRefExoticComponent<
+    PropsWithoutRef<PickPartial<ComponentProps<T>, 'children'>> & RefAttributes<unknown>
+  > & {
+    Screen: (props: ScreenProps<TOptions, TState, TEventMap>) => null;
+  };
 }