+++ packages/expo-router/src/typed-routes/generate.ts
@@ -1,33 +1,117 @@
 import { RouteNode } from '../Route';
 import { getRoutes } from '../getRoutes';
-import { isTypedRoute, removeSupportedExtensions } from '../matchers';
+import { removeSupportedExtensions } from '../matchers';
 import { RequireContext } from '../types';
 
 // /[...param1]/ - Match [...param1]
 const CATCH_ALL = /\[\.\.\..+?\]/g;
 // /[param1] - Match [param1]
 const SLUG = /\[.+?\]/g;
+// /(group)/path/(group2)/route - Match [(group), (group2)]
+const GROUP = /(?:^|\/)\(.*?\)/g;
 
-export function getTypedRoutesDeclarationFile(ctx: RequireContext) {
-  const staticRoutes = new Set<string>();
-  const dynamicRoutes = new Set<string>();
-  const dynamicRouteContextKeys = new Set<string>();
+export type GetTypedRoutesDeclarationFileOptions = {
+  partialTypedGroups?: boolean;
+  testIgnoreComments?: boolean;
+};
 
+export function getTypedRoutesDeclarationFile(
+  ctx: RequireContext,
+  {
+    partialTypedGroups = false,
+    testIgnoreComments = false,
+  }: GetTypedRoutesDeclarationFileOptions = {}
+) {
   let routeNode: RouteNode | null = null;
 
   try {
     routeNode = getRoutes(ctx, {
+      ignore: [/_layout\.[tj]sx?$/], // Skip layout files
       platformRoutes: false, // We don't need to generate platform specific routes
+      notFound: false, // We don't need +not-found routes either
       ignoreEntryPoints: true,
       ignoreRequireErrors: true,
-      importMode: 'async',
+      importMode: 'async', // Don't load the file
     });
   } catch {
     // Ignore errors from `getRoutes`. This is also called inside the app, which has
     // a nicer UX for showing error messages
   }
 
-  walkRouteNode(routeNode, '', staticRoutes, dynamicRoutes, dynamicRouteContextKeys);
+  const groupedNodes = groupRouteNodes(routeNode);
+  const staticRoutesStrings: string[] = ['Router.RelativePathString', 'Router.ExternalPathString'];
+  const staticRouteInputObjects: string[] = [
+    '{ pathname: Router.RelativePathString, params?: Router.UnknownInputParams }',
+    '{ pathname: Router.ExternalPathString, params?: Router.UnknownInputParams }',
+  ];
+  const staticRouteOutputObjects: string[] = [
+    '{ pathname: Router.RelativePathString, params?: Router.UnknownOutputParams }',
+    '{ pathname: Router.ExternalPathString, params?: Router.UnknownOutputParams }',
+  ];
+
+  for (const type of groupedNodes.static) {
+    staticRoutesStrings.push(
+      contextKeyToType(type + "${`?${string}` | `#${string}` | ''}", partialTypedGroups)
+    );
+    staticRouteInputObjects.push(
+      `{ pathname: ${contextKeyToType(type, partialTypedGroups)}; params?: Router.UnknownInputParams; }`
+    );
+    staticRouteOutputObjects.push(
+      `{ pathname: ${contextKeyToType(type, partialTypedGroups)}; params?: Router.UnknownOutputParams; }`
+    );
+  }
+
+  const dynamicRouteStrings: string[] = [];
+  const dynamicRouteInputObjects: string[] = [];
+  const dynamicRouteOutputObjects: string[] = [];
+
+  for (const [dynamicRouteTemplate, paramsNames] of groupedNodes.dynamic) {
+    const inputParams = paramsNames
+      .map((param) => {
+        const key = param.startsWith('...') ? param.slice(3) : param;
+        const value = param.startsWith('...') ? '(string | number)[]' : 'string | number';
+        return `${key}: ${value};`;
+      })
+      .join('');
+
+    const outputParams = paramsNames
+      .map((param) => {
+        const key = param.startsWith('...') ? param.slice(3) : param;
+        const value = param.startsWith('...') ? 'string[]' : 'string';
+        return `${key}: ${value};`;
+      })
+      .join('');
+
+    dynamicRouteStrings.push(
+      contextKeyToType(
+        dynamicRouteTemplate
+          .replaceAll(CATCH_ALL, '${string}')
+          .replaceAll(SLUG, '${Router.SingleRoutePart<T>}'),
+        partialTypedGroups
+      )
+    );
+
+    dynamicRouteInputObjects.push(
+      `{ pathname: ${contextKeyToType(dynamicRouteTemplate, partialTypedGroups)}, params: Router.UnknownInputParams & { ${inputParams} } }`
+    );
+    dynamicRouteOutputObjects.push(
+      `{ pathname: ${contextKeyToType(dynamicRouteTemplate, partialTypedGroups)}, params: Router.UnknownOutputParams & { ${outputParams} } }`
+    );
+  }
+
+  const href = [
+    ...staticRoutesStrings,
+    ...staticRouteInputObjects,
+    ...dynamicRouteStrings,
+    ...dynamicRouteInputObjects,
+  ].join(' | ');
+
+  const hrefInputParams = [...staticRouteInputObjects, ...dynamicRouteInputObjects].join(' | ');
+  const hrefOutputParams = [...staticRouteOutputObjects, ...dynamicRouteOutputObjects].join(' | ');
+
+  const tsExpectError = testIgnoreComments
+    ? '// @ts-ignore-error -- During tests we need to ignore the "duplicate" declaration error, as multiple fixture declare types \n      '
+    : '';
 
   return `/* eslint-disable */
 import * as Router from 'expo-router';
@@ -36,99 +120,106 @@ export * from 'expo-router';
 
 declare module 'expo-router' {
   export namespace ExpoRouter {
-    export interface __routes<T extends string = string> extends Record<string, unknown> {
-      StaticRoutes: ${setToUnionType(staticRoutes)};
-      DynamicRoutes: ${setToUnionType(dynamicRoutes)};
-      DynamicRouteTemplate: ${setToUnionType(dynamicRouteContextKeys)};
+    export interface __routes<T extends string | object = string> {
+      ${tsExpectError}hrefInputParams: ${hrefInputParams};
+      ${tsExpectError}hrefOutputParams: ${hrefOutputParams};
+      ${tsExpectError}href: ${href};
     }
   }
 }
 `;
 }
 
-/**
- * Walks a RouteNode tree and adds the routes to the provided sets
- */
-function walkRouteNode(
+function groupRouteNodes(
   routeNode: RouteNode | null,
-  parentRoutePath: string,
-  staticRoutes: Set<string>,
-  dynamicRoutes: Set<string>,
-  dynamicRouteContextKeys: Set<string>
-) {
-  if (!routeNode) return;
-
-  addRouteNode(routeNode, parentRoutePath, staticRoutes, dynamicRoutes, dynamicRouteContextKeys);
-
-  parentRoutePath = `${removeSupportedExtensions(`${parentRoutePath}/${routeNode.route}`).replace(/\/?index$/, '')}`; // replace /index with /
-
-  for (const child of routeNode.children) {
-    walkRouteNode(child, parentRoutePath, staticRoutes, dynamicRoutes, dynamicRouteContextKeys);
+  groupedContextKeys = {
+    static: new Set<string>(),
+    dynamic: new Map<string, string[]>(),
   }
-}
-
-/**
- * Given a RouteNode, adds the route to the correct sets
- * Modifies the RouteNode.route to be a typed-route string
- */
-function addRouteNode(
-  routeNode: RouteNode | null,
-  parentRoutePath: string,
-  staticRoutes: Set<string>,
-  dynamicRoutes: Set<string>,
-  dynamicRouteContextKeys: Set<string>
 ) {
-  if (!routeNode?.route) return;
-  if (!isTypedRoute(routeNode.route)) return;
+  if (!routeNode) {
+    return groupedContextKeys;
+  }
 
-  let routePath = `${parentRoutePath}/${removeSupportedExtensions(routeNode.route).replace(/\/?index$/, '')}`; // replace /index with /
+  // Skip non-route files
+  if (routeNode.type !== 'route') {
+    // Except the root layout
+    if (routeNode.route === '') {
+      for (const child of routeNode.children) {
+        groupRouteNodes(child, groupedContextKeys);
+      }
+      return groupedContextKeys;
+    }
 
-  if (!routePath.startsWith('/')) {
-    routePath = `/${routePath}`;
+    return groupedContextKeys;
+  }
+
+  let routeKey: string;
+
+  if (routeNode.generated) {
+    // Some routes like the root _layout, _sitemap, +not-found are generated.
+    // We cannot use the contextKey, as their context key does not specify a route
+    routeKey = routeNode.route;
+  } else {
+    routeKey = removeSupportedExtensions(routeNode.contextKey)
+      .replace(/\/index$/, '') // Remove any trailing /index
+      .replace(/^\./, ''); // Remove any leading .
+  }
+
+  routeKey ||= '/'; // A routeKey may be empty for contextKey '' or './index.js'
+
+  if (!routeKey.startsWith('/')) {
+    // Not all generated files will have the `/` prefix
+    routeKey = `/${routeKey}`;
   }
 
   if (routeNode.dynamic) {
-    for (const path of generateCombinations(routePath)) {
-      dynamicRouteContextKeys.add(path);
-      dynamicRoutes.add(
-        `${path.replaceAll(CATCH_ALL, '${string}').replaceAll(SLUG, '${Router.SingleRoutePart<T>}')}`
-      );
-    }
+    groupedContextKeys.dynamic.set(
+      routeKey,
+      routeKey
+        .split('/')
+        .filter((segment) => {
+          return segment.startsWith('[') && segment.endsWith(']');
+        })
+        .map((segment) => {
+          return segment.slice(1, -1);
+        })
+    );
   } else {
-    for (const combination of generateCombinations(routePath)) {
-      staticRoutes.add(combination);
-    }
+    groupedContextKeys.static.add(routeKey);
   }
+
+  for (const child of routeNode.children) {
+    groupRouteNodes(child, groupedContextKeys);
+  }
+
+  return groupedContextKeys;
 }
 
-/**
- * Converts a Set to a TypeScript union type
- */
-const setToUnionType = <T>(set: Set<T>) => {
-  return set.size > 0
-    ? [...set]
-        .sort()
-        .map((s) => `\`${s}\``)
-        .join(' | ')
-    : 'never';
-};
-
-function generateCombinations(pathname) {
-  const groups = pathname.split('/').filter((part) => part.startsWith('(') && part.endsWith(')'));
-  const combinations: string[] = [];
-
-  function generate(currentIndex, currentPath) {
-    if (currentIndex === groups.length) {
-      combinations.push(currentPath.replace(/\/{2,}/g, '/'));
-      return;
+function contextKeyToType(contextKey: string, partialTypedGroups: boolean) {
+  // If the route has groups, turn them into template strings
+  const typeWithGroups = contextKey.replaceAll(GROUP, (match) => {
+    const groups = match.slice(2, -1); // Remove the leading ( and the trailing )
+    // When `partialRoutes` is enabled, we always change a group to a template
+    if (groups.length > 1 || partialTypedGroups) {
+      // Ensure each group has the trailing slash
+      const groupsAsType = groups.split(',').map((group) => `'/(${group})'`);
+      // `partialRoutes` allow you to skip a group
+      if (partialTypedGroups) {
+        groupsAsType.push("''");
+      }
+      // Combine together into a union
+      return `\${${groupsAsType.join(' | ')}}`;
+    } else {
+      return match;
     }
+  });
 
-    const group = groups[currentIndex];
-    const withoutGroup = currentPath.replace(`/${group}`, '');
-    generate(currentIndex + 1, withoutGroup);
-    generate(currentIndex + 1, currentPath);
+  const typeWithoutGroups = contextKey.replaceAll(GROUP, '');
+
+  if (typeWithGroups === typeWithoutGroups) {
+    return `\`${typeWithGroups}\``;
+  } else {
+    return `\`${typeWithGroups}\` | \`${typeWithoutGroups}\``;
   }
-
-  generate(0, pathname);
-  return combinations;
 }