+++ packages/expo-updates/e2e/setup/project.ts
@@ -29,6 +29,7 @@ function getExpoDependencyChunks({
     ['@expo/cli', 'expo', 'expo-asset', 'expo-modules-autolinking'],
     ['expo-manifests'],
     ['@expo/prebuild-config', '@expo/metro-config', 'expo-constants'],
+    ['@expo/image-utils'],
     [
       'babel-preset-expo',
       'expo-application',