+++ packages/babel-preset-expo/CHANGELOG.md
@@ -10,6 +10,20 @@
 
 ### 💡 Others
 
+- Added EAS Updates support for DOM Components. ([#32502](https://github.com/expo/expo/pull/32502) by [@kudo](https://github.com/kudo))
+
+## 12.0.0-preview.4 — 2024-10-31
+
+### 🐛 Bug fixes
+
+- Don't assert `client-only` in SSR bundles. ([#32479](https://github.com/expo/expo/pull/32479) by [@EvanBacon](https://github.com/EvanBacon))
+
+## 12.0.0-preview.3 — 2024-10-30
+
+### 🎉 New features
+
+- Add support for server actions that use `export default`. ([#32458](https://github.com/expo/expo/pull/32458) by [@EvanBacon](https://github.com/EvanBacon))
+
 ## 12.0.0-preview.2 — 2024-10-29
 
 ### 🎉 New features